Changed router behaviour to add a priority for hardcoded routes.
This changes the "linear" behaviour of the router with a somewhat smarter one. The routes are now sorted to prioritise hardcoded values. This means that if a URI with the value "/test" is specified, it will be matched before "/{id}" even if "/{id}" is declared before "/test" is the execution order.